move reusable utils to dedicated module
